use history hook in the image context
